hydra-node/requirements.txt
requests==2.31.0
cbor2==5.4.6
base58==2.1.1
PyNaCl==1.5.0
python-dateutil==2.8.2
flask==2.3.3
